feat: Implement a new terminal extension with backend PTY handling, frontend UI components, and comprehensive tests for its capabilities and payloads.
